src/requirements.txt
Django==1.10
psycopg2==2.6.1
wheel==0.24.0

django-bootstrap3==6.2.2
django-adminplus==0.5
django-model-utils==2.4
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
django-rest-marshmallow==1.0.0
django-crispy-forms==1.6.0
django-filter==0.13.0
django-guardian==1.4.3
django-cors-headers

Markdown==2.6.6
whitenoise==3.0
marshmallow==2.7.3
PyJWT==1.4.0
PyYAML==3.11
django-nested-admin==3.0.8
django-rest-swagger==0.3.6
WeasyPrint==0.30

# our packages are only listed here for information purpose
# - on dev box, they should be installed from tarballs generated from source code
# - on prod box, they are installed from tarballs bundled together
# grid-core-db==0.2

